fix bug where SPI was sending bogus data
It appears that sometimes when the SPI 'resends' the same byte, or has resent the same byte, the next byte to be sent gets mangled. For example, if the 'fill'/no-data byte is 0xFF, sometimes the length byte picks up the MSB and 10 becomes 138. Also, sometimes the length byte is dropped altogether. The new implementation runs the TXE interrupt continuously, refilling the transmit buffer with the fill byte (0x00) if there is no data ready to be sent.